Revert "hardware: add 64bit support [1/1]"
This reverts commit 76b93cabc61e9fa93f2498898165958b9233d844.

+++ LibAudio/amadec/adec_read.c
@@ -111,7 +111,7 @@ int uio_init(aml_audio_dec_t *audec)
     }
     if (phys_offset == 0)
         phys_offset = ((AIU_AIFIFO_CTRL + addr_offset) << 2) & (pagesize - 1);
-    reg_base = (volatile unsigned*)((unsigned long)memmap + phys_offset);
+    reg_base = (volatile unsigned*)((unsigned)memmap + phys_offset);
     pthread_mutex_unlock(&uio_mutex);
 
     return 0;
